added: CSocketServer
git-svn-id: https://subversion.gsi.de/dgrid/D-Grid/Include@518 47ffe4a8-6206-0410-815e-8f101de37557

+++ ErrorCode.h
@@ -18,7 +18,7 @@
 namespace MiscCommon
 {
 
-    typedef size_t ERRORCODE;
+    typedef unsigned int ERRORCODE;
     typedef const ERRORCODE ERRORCODE_C;
